refactor(symbols): move find_symbol + SCIP helpers into pipeline (arch PR C)
PR-C of the SymbolQuery pipeline series. `find_symbol` joins
`get_ranked_context` behind `SymbolQueryPipeline`. handlers.rs's
`find_symbol` entry collapses from a 290-line orchestration to a
3-line dispatch into `SymbolQueryRequest::FindSymbol`.

What moved into `symbol_query/find_symbol.rs`:
  - `run_find_symbol`: the orchestration body itself
    (argument parsing + SCIP precise lookup with stale-index meta
    degradation + tree-sitter fallback with 0-result fallback_hint).
  - SCIP enrichment helpers (5): `heuristic_body_slice`,
    `read_signature_line`, `looks_like_signature`,
    `humanize_scip_name_path`, plus `HEURISTIC_BODY_LINES`.
  - 4 unit test modules — `find_symbol_argument_tests`,
    `looks_like_signature_tests`, `humanize_scip_name_path_tests`,
    `read_signature_line_tests`. The argument tests now drive
    `SymbolQueryPipeline::run(SymbolQueryRequest::FindSymbol { ... })`
    to exercise the actual dispatch entry.

All five SCIP helpers are now `fn` (private) inside the pipeline
module — they were previously `pub(super)` in handlers.rs and existed
only because find_symbol used them.

Visibility shifts:
  - `symbols::formatter::compact_symbol_bodies`: pub(super) →
    pub(crate) (find_symbol pipeline calls it). PR-E will tighten.
  - `symbols::formatter`: mod → pub(crate) mod (same reason).

Deletion test: removing `tools/symbol_query/find_symbol.rs` would
re-disperse 5 SCIP helpers + 290-line orchestration back into
handlers.rs + force the unit tests to inline-import private helpers.
The find_symbol semantics concentrate in one file.

Net diff: +662/-646 LOC across 5 files. handlers.rs shrinks
1337 → 707 LOC (47% smaller).

Build matrix:
  - cargo check --workspace --features scip-backend,semantic ✓ (1.07s)
  - cargo check --workspace --no-default-features ✓
  - cargo clippy --features semantic -- -D warnings ✓
  - cargo clippy --no-default-features -- -D warnings ✓
  - cargo fmt --all -- --check ✓
  - cargo test -p codelens-mcp --bin codelens-mcp --features semantic
    → 634 passed / 1 ignored / 0 failed (61.32s)
