+++ crates/codelens-mcp/src/tools/symbol_query/mod.rs
@@ -1,9 +1,9 @@
 //! SymbolQueryPipeline — the single seam through which every
 //! symbol-shape tool will execute.
 //!
-//! Today only `RankedContext` lives behind the seam (PR-B). PR-C
-//! adds `FindSymbol`; PR-D adds `SymbolsOverview`. Once those land,
-//! the symbol-shape tool surface boils down to argument parsing in
+//! `RankedContext` landed in PR-B; `FindSymbol` lands in PR-C (this
+//! PR); `SymbolsOverview` lands in PR-D. Once PR-D ships, the
+//! symbol-shape tool surface boils down to argument parsing in
 //! `tools/symbols/handlers.rs` followed by a single call to
 //! `SymbolQueryPipeline::run`, and the whole semantics of any one
 //! tool can be read inside this module.
@@ -24,8 +24,10 @@ use super::AppState;
 use crate::tool_runtime::ToolResult;
 use serde_json::Value;
 
+mod find_symbol;
 mod ranked_context;
 
+pub(crate) use find_symbol::run_find_symbol;
 pub(crate) use ranked_context::run_ranked_context;
 
 pub struct SymbolQueryPipeline<'s> {
@@ -33,10 +35,10 @@ pub struct SymbolQueryPipeline<'s> {
 }
 
 /// One variant per symbol-shape MCP tool. Each variant captures the
-/// raw JSON `arguments` for now; PR-C/D may parse them into typed
-/// inner structs once a second variant exists to share the shape.
+/// raw JSON `arguments`; PR-D will add `SymbolsOverview`.
 pub enum SymbolQueryRequest<'a> {
     RankedContext { arguments: &'a Value },
+    FindSymbol { arguments: &'a Value },
 }
 
 impl<'s> SymbolQueryPipeline<'s> {
@@ -49,6 +51,7 @@ impl<'s> SymbolQueryPipeline<'s> {
             SymbolQueryRequest::RankedContext { arguments } => {
                 run_ranked_context(self.state, arguments)
             }
+            SymbolQueryRequest::FindSymbol { arguments } => run_find_symbol(self.state, arguments),
         }
     }
 }
